requirements.txt
beautifulsoup4==4.11.1
biopython==1.79
fastapi==0.85.1
Flask==2.2.2
gffutils==0.11.1
numpy==1.23.4
oauth2client==4.1.3
openpyxl==3.0.10
pandas==1.5.1
pybedtools==0.9.0
pydantic==1.10.2
pydantic-webargs==1.1.0
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
validators==0.20.0
wget==3.2